test: 미션2 데이터베이스 연동 테스트 코드 추가 #step4
